requirements.txt
pytest==7.3.1
azure-functions==1.14.0
neobase==0.4.0